+++ packages/lab/src/index.ts
@@ -31,6 +31,7 @@ export {
   StaticInputAdornment,
 } from "./input-legacy";
 export * from "./layer-layout";
+export * from "./link-button";
 export * from "./list";
 export type {
   ListChangeHandler as ListChangeHandlerDeprecated,