+++ site/src/examples/carousel/Card.tsx
@@ -41,6 +41,7 @@ export const Card = (): ReactElement => {
                 media={
                   <img
                     aria-hidden={true}
+                    alt=""
                     className={styles.carouselImage}
                     src={slide.image}
                   />