+++ site/src/examples/carousel/MultipleSlides.tsx
@@ -53,6 +53,7 @@ export const MultipleSlides = (): ReactElement => {
                 media={
                   <img
                     aria-hidden={true}
+                    alt=""
                     className={styles.carouselImage}
                     src={slide.image}
                   />