+++ site/src/examples/link-button/ExternalLink.tsx
@@ -1,0 +1,12 @@
+import { LinkButton } from "@salt-ds/lab";
+import type { ReactElement } from "react";
+
+export const ExternalLink = (): ReactElement => (
+  <LinkButton
+    href="https://www.saltdesignsystem.com"
+    rel="noopener"
+    target="_blank"
+  >
+    Visit Salt
+  </LinkButton>
+);