+++ site/src/examples/scrim/WithChild.tsx
@@ -40,6 +40,7 @@ export const WithChild = (): ReactElement => {
                       fill: "var(--salt-sentiment-positive-foreground-decorative)",
                     }}
                     size={1}
+                    aria-hidden
                   />
                 </Display3>
                 <Text
@@ -60,6 +61,7 @@ export const WithChild = (): ReactElement => {
                       fill: "var(--salt-sentiment-negative-foreground-decorative)",
                     }}
                     size={1}
+                    aria-hidden
                   />
                 </Display3>
                 <Text
@@ -82,6 +84,7 @@ export const WithChild = (): ReactElement => {
                       fill: "var(--salt-sentiment-positive-foreground-decorative)",
                     }}
                     size={1}
+                    aria-hidden
                   />
                 </Display3>
                 <Text
@@ -102,6 +105,7 @@ export const WithChild = (): ReactElement => {
                       fill: "var(--salt-sentiment-positive-foreground-decorative)",
                     }}
                     size={1}
+                    aria-hidden
                   />
                 </Display3>
                 <Text