+++ site/src/examples/tag/WithIcon.tsx
@@ -1,15 +1,15 @@
 import { FlowLayout, Tag } from "@salt-ds/core";
-import { ClockIcon, LineChartIcon } from "@salt-ds/icons";
+import { ChartLineIcon, ClockIcon } from "@salt-ds/icons";
 import type { ReactElement } from "react";
 
 export const WithIcon = (): ReactElement => (
   <FlowLayout>
     <Tag>
-      <LineChartIcon />
+      <ChartLineIcon aria-hidden />
       Data
     </Tag>
     <Tag variant="secondary">
-      <ClockIcon /> Coming soon
+      <ClockIcon aria-hidden /> Coming soon
     </Tag>
   </FlowLayout>
 );